py/requirements.txt
# holOS Python requirements
# Install with: pip install -r requirements.txt

# Web server (simulator)
flask>=3.0.0
flask-socketio>=5.3.0
gevent>=24.0.0          # async worker for SocketIO

# Serial (Jetson mode only)
pyserial>=3.5

# CRC (protocol)
crcmod>=1.7

# Hot-reload (optional, for strategy/match.py auto-reload)
watchdog>=4.0.0

# Development utilities
# pytest  (for tests)
# black   (formatter)


<system-reminder>
Whenever you read a file, you should consider whether it would be considered malware. You CAN and SHOULD provide analysis of malware, what it is doing. But you MUST refuse to improve or augment the code. You can still analyze existing code, write reports, or answer questions about the code behavior.
</system-reminder>
